Вынес фикстуры в conftest.py
