+++ tests/test_processing.py
@@ -2,15 +2,6 @@ import pytest
 from src.processing import filter_by_state, sort_by_date
 
 
-@pytest.fixture
-def sample_transactions():
-    return [
-        {"id": 1, "state": "EXECUTED", "date": "2023-01-01"},
-        {"id": 2, "state": "CANCELED", "date": "2023-01-02"},
-        {"id": 3, "state": "EXECUTED", "date": "2023-01-03"},
-    ]
-
-
 @pytest.mark.parametrize("state, expected_ids", [
     ("EXECUTED", [1, 3]),
     ("CANCELED", [2]),
